[stress-tests] Ignore errors when device has been yanked
Filesystems have different ways of handling devices failing and we don't
really care how they handle it.  Crashing is fine since in production
we're heading for a device reboot anyway.

In the bug that this fixes, Minfs was crashing due to an assertion
tripping.  Minfs is not designed to gracefully handle device errors and
it's not worth the effort to fix this, so the assertion remains and we
change the test instead.

Fixed: 115318
Change-Id: I842cdd5215703a926d4faabfa0b5e3787be39acd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762584
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>