[assembly] Make update epoch configurable
The epoch used in the update package can now be configured by the
product/board.

We have a platform epoch that is added to the product epoch in order to
ensure that the platform can bump the epoch for all product if
necessary.

Fixed: 109440
Change-Id: I2caeb4d5c6d0fc88d59263bdca127cbbea95902c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762083
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>