[fidl] test unknown_interaction_migration
Enable the unknown_interaction_migration experimental flag for all the
FIDL test files in tools/fidl/fidlc/testdata. This should demonstrate
that enabling this flag does not affect the generated code for existing
libraries.

Skimming the golden files, there appear to be a few changes:
- JSON IR files now include openness modifiers and strictness modifiers
  for methods. This is expected and is correct as long as they are
  always "closed" and "strict".
- Generated rust code includes a comment stating which experiments are
  enabled. This has no effect on behavior.

These should be the only changes to the goldens.

I am not adding the explicit openness and strictness modifiers yet
because I want to simulate the release process; when the
unknown_interaction_migration experiment gets enabled, I will not be
atomically changing all other FIDL files at the same time, so it is good
to demonstrate its behavior here. Additionally, these files can serve as
a test case for automated tooling to apply the closed and strict
modifiers later, which will be needed to cover the number of FIDL files
which will need to be changed.

Bug: 88366
Change-Id: I61f67c5b188c83c8ebfb263d6d36930cdc5dba08
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762188
Commit-Queue: Zachary Stewart <zstewart@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>