[cm_config][compile] Enforce config field inclusion at build time
Fixes: when a developer adds a new field to CM's config json file, it is
easy to overlook that the field must also be added to the `extend`
function in order for the config compiler to set the field in the
compiled Config struct.

This CL makes it so the rust compiler will enforce that new fields must
be included in this function. Also renames `extend` to `merge`, since a
new Config is being produced.

Change-Id: Ie418514ad404eadb6787536ca6b30e721760cc1e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758609
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>