[memory] Trace memory allocation and deallocation.
This investigation tool for memory misbehavior intercepts all
allocations and deallocation within a process linked with
`libmemory_trace.so`.

Each event is logged to the Trace API so that it can be collected on the
developper's workstation with standard `ffx trace` tool.

`fxt_to_pprof` is additional host tool matches allocation and
deallocation, identify the memory that remains allocated an the end of
the profile session.  The result is exported to pprof format so that it
can be symbolized and presented by pprof.

More information in README.md.

Bug: 111236
Change-Id: I7c00e184377ff5b075abf34f77e4bee1c523a305
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704308
Reviewed-by: Guillaume Chelfi <chelfi@google.com>
Commit-Queue: Pierre Labatut <plabatut@google.com>
Reviewed-by: Étienne J. Membrives <etiennej@google.com>