[pbm] Add --force-repo flag to product-bundle get
The attached bug requests a way for users to overwrite a repository when
downloading a product bundle, as was the behavior before
https://fuchsia-review.git.corp.google.com/c/fuchsia/+/735685.

This CL adds a new flag, --force-repo, which allows the product-bundle
get command to proceed even if a repository with the same name already
exists. The existing repo registry entry will be overwritten with the
path for the bundle being downloaded. Note that even if the bundle is
already in the local cache, the repo will be overwritten, but the bundle
will not be re-downloaded unless the user also includes the --force
flag.

Bug: 112892
Change-Id: I983a7a3d87ef86fb04a057371bee77ab578483c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757769
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>