[build] Add deny_warnings flag.
This flag is enabled by default, and promotes Rust warnings
to errors, preserving the existing behavior. In the future
this can be extended to other languages, including C++.

This is intended as a replacement for the rust_cap_lints arg
which is used to achieve similar development velocity goals
but which is very coarse-grained. If there are warning-by-default
lints we want to still be errors for local development they will
still be demoted back to warnings by cap_lints, which can produce
confusion when e.g. an unused Result is causing a test failure
but isn't producing an error. By controlling whether -Dwarnings
is passed to rustc at all, we can maintain granular control over
disallowed lints during local development.

This implementation also selects several key lints which should
always be errors, even when the flag is disabled during local
development.

Change-Id: Ie23000b40e344374c3f7b041e36b37863fbd6ff5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762328
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Joseph Ryan <josephry@google.com>
Commit-Queue: Adam Perry <adamperry@google.com>