[hwinfo] Log request stream handler errors instead of panicking.
When clients drop connections to hwinfo without awaiting a response, the
server may receive a PEER_CLOSED error. Since this is a Result::Err,
this propagates up to an expect() call that causes a panic and crashes
the component.

Instead of panicking, these errors are logged and the task that
services the client is allowed to end.

Bug: b/257979013
Bug: 110564
Test: CQ
Change-Id: I97f7c1e648f31e8129b44a13fd838c33351ff5ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761963
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>
Commit-Queue: Michael Brunson <mbrunson@google.com>