[debugger] Resolve member types in pretty-printers.
Allow pretty-printing code to refer to member types on the object being
pretty-printed as if the code was running in an object member function.
Previously only variable names were resolved.

Adds a DWARF ULEB number encoder. The unit test for this exposed a bug
in the decoder for numbers greater than 32-bits. Fix this.

Change-Id: I8a6a571328039e00296baacd88000c1a06ee36d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761765
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>