[assembly] Sort shell command entries in the AIB manifest
Sort the list of shell commands by package-name in the JSON manifest
for each AIB when it's created.  This enables diff'ing after
making changes to the order that items are presented for adding to
the AIB, without changing the order the of items in the file.

Change-Id: Icc4db5ae39c62e9c78da8f3b9f72d24083fb5ad0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762983
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>