[intel-i915-tgl] Fix error message when there is no DP sink detected.
It is possible that a DPRX (receiver) device returns a value
of 0 for sink count, which indicates that the repeater / dock
does not have any sink (display) device connected.

This change fixes the error message so that we can distinguish
if a DP port has no sinks or has multiple sinks (MST).

Bug: 103986
Test: HDMI port 1 (a DP repeater on DDI TC 1) on Intel NUC11TNHi5.
Test: Added test case DpDisplayTest.NoSinkNotSupported
      to explicitly test the zero-sink case.
Change-Id: I72e58aad8e4a0dc4b5915df2c8df51af835dbfa3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763404
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>