[build] Add //tools/docsgen/tests for host tests
This CL splits //tools/docsgen/BUILD.gn host tests off into a
subdirectory, allowing us to assert on the toolchain as recommended in
https://fuchsia.dev/fuchsia-src/development/build/build_system/best_practices_toolchains?hl=en#assert-on-the-expected-toolchain.

It also fixes the test which wasn't working before (it wasn't reachable
in the build graph).

The motivation for this change: this build file was causing problems for
removing the FIDL toolchain (fxbug.dev/108355).

Test: fx test //tools/docsgen/tests
Bug: 108355
Change-Id: I41a46193fd1513ecd872250cc7bbdcde1938a626
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762923
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Theodore Siu <theosiu@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>