[fx][serve] Don't error out if server is using port 0
If the repository server is configured to run on port 0, it will be
allocated a random port. We don't have an easy way to determine which
port the server is running on, so emit a warning in this situation
instead of failing.

Bug: 115559
Change-Id: I107825834307c6d73e698241524b21ea151f7ede
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762936
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>