[audio] Implement OutputDevicePipeline
This is a reimplementation of audio_core/v2/output_device_pipeline.cc.
This class manages a collection of graph nodes that provide access to an
output device. The device is represented by a ConsumerNode, and above
that consumer is a tree of MixerNodes, CustomNodes, and at most one
SplitterNode (which represents a loopback interface). This tree is
defined by audio_core's PipelineConfig.

Each MixerNode manages one or more "usages". ProducerNodes (a.k.a. renders)
will connect to an output pipeline via the MixerNode which managers that
renderer's usage.

Bug: 98652
Change-Id: I7da9dcb4d743fbb0b229c30cc1b226e921d2a5ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761112
Reviewed-by: Alper Gungormusler <alperg@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>