[early-boot-instrumentation] Add more logging
Attempting to identify the source of the corruption in some
cases in coverage builds.

We will log every time we encounter a stashed svc handle,
every time we encounter a debugdata.Publish request,
every time we extract a vmo to be published.

Also every time we see a kernel and/or physboot entry.

No functional changes.

Bug: 114996
Test: fx build
Change-Id: I9ed695702dc23ea047549abe85f810045017128b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762072
Reviewed-by: Gulfem Savrun Yeniceri <gulfem@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>