[kernel_zbi_test] Introduce cuckoo_kernel_zbi_test
This templates takes a zbi input that is used as the kernel zbi. The
only requirement for the kernel zbi is that it must reach userboot at
some point. The correct set of arguments are added such that component
manager is always launched at the end (e.g. core-tests) and proceed
as usual.

If the zbi image has a custom name an override can be provided through
the template, this allows assembled_system to select the right image
for the final assembly.

Bug: 109878
Test: manually verifies system boots and interacts with it through tools
Change-Id: I48198ef2ea1f9871cb2bb94997939749a55d3def
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711799
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>