[audio] Support another legacy FIDL format
There are three "audio format" definitions in FIDL:

1. fuchsia.audio.Format
2. fuchsia.mediastreams.AudioFormat
3. fuchsia.media.AudioStreamType

(1) is the new format (not yet released to the SDK). (3) is the original
format (released to the SDK). (2) was an in-between step that was
released to the SDK before (1) was finalized.

Currently we have Format::Create* functions for (1) and (2). This CL
adds a function for (3).

Bug: 98652
Change-Id: I7a7027a541d333aed65a71858448bf30078e10d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761499
Commit-Queue: Tom Bergan <tombergan@google.com>
Reviewed-by: Alper Gungormusler <alperg@google.com>