[fidl][cpp] Declare IsProtocol in markers
Previously, we generate code in the messaging layer to teach the
compiler that a FIDL marker type is a protocol (fidl::IsProtocol).
That's too late because code in the domain objects layer may want to
assert that a marker type is a protocol, to catch programming errors.

Also changes the fidlgen_cpp library to omit banjo and syscall protocols
from the compiled C++ IR, because we are not prepared to generate code
for those protocols at this point (this may change in the future).

Change-Id: I410f519741c7dfff541fd6d8959147d3610ea573
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763222
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>