[fidl][cpp] Remove MessageTraits
This is in preparation of changing distinct message wrapper types
such as fidl::Event<M> to simply aliases. Because they won't be separate
types, we cannot template definitions over them.

It is mostly a pure refactor. The NaturalMessageConverter mechanism was
meant to generically convert domain objects. But it turned out to be an
over generalization, and only used to fold errors in two way results. In
removing its dependency on MessageTraits we can specialize it and
simplify things. In doing so we also remove special transformations for
error syntax events, since those are deprecated (fxbug.dev/98319).

Bug: 101525
Bug: 98319

Change-Id: Id9df993f9d6be98dc10bab8d6c714b9bf6cae4ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763382
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Zachary Stewart <zstewart@google.com>