[vsh] Unconditionally try to launch termina
In certain situations the VM may boot but the container is not properly
initialized. Subsequent calls to `guest vsh -c` would see that the VM is
running and skip calling termina::launch not realizing that the
container is not running, leading to an unrecoverable failure.

Since StartAndGetLinuxGuestInfo is cognizant of the state of both the VM
and container and will retry or be a no-op in the appropriate situations
we should just unconditionally call it instead.

Change-Id: Icdcb76b62a10c03a6cb183f60f87f1be43280c3d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/763085
Fuchsia-Auto-Submit: Jonathan Bailey <jonathanbailey@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>