[driver_manager] Have composite fragments use FidlProxies
Composite fragments now use the FidlProxy flow to do banjo proxying.

When a composite device needs to do banjo proxying for a parent device
it will:

1) Bind fragment.so to the parent device
2) Create a FidlProxy in the composite's driver host.
3) Bind fragment.proxy.so to the FidlProxy
4) Create the composite device

So now the banjo proying is done by fragment.so and fragment.proxy.so,
and these drivers use FidlProxy to do this.

This gets us a step closer to deleting the non FidlProxy
implementations.

Bug: 114085
Multiply: libdriver-integration-test

Change-Id: Ifbec12bab5a416b925f736e8b1c530fa520a8ed3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757587
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>