[test_manager] Instantiate test in collection
Change the test manager realm setup to launch the test root component in
a collection, instead of making it a static child.

This is to prepare for a mitigation of fxbug.dev/76579 which will
require destroying the test component independently of the entire realm
(see stacked change).

The test component is now instantiated as lazy instead of eager (dynamic
components don't support eager). A change to pkgdir-integration-tests
was required to make it compatible with this behavior.

Run-All-Tests: true

Bug: 76579

Change-Id: I0837703581c5fadfaebfa24e05acaee0e289b6ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762300
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Gary Bressler <geb@google.com>