[driver-index] Use separate MatchDriverArgs
Use a separate table for matching a driver.
This is in preparation for adding more things to the
NodeAddArgs.

It also turns out that symbols and offers are not being
used to do matching, so this cleans up the API a bit.

Bug: 114993

Change-Id: I6bf55c52de1a7ee9544b148e1d32da2ed96c5b8b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764313
Reviewed-by: Christopher Anderson <cja@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
API-Review: Christopher Anderson <cja@google.com>