[botanist] Parse flags with forward compatibility
To be able to change the recipes code a way of parsing the flags
that is both backwards and forward compatible is necessary.

Most of the code in this CL is temporary and will be remove in a
future CL once the recipe code that references botanist and testrunner
gets changed.

Bug: 111922
Change-Id: Id021b71ba3aa2ce1b306da298ffe2115112ac350
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/764664
Commit-Queue: Andrés Villegas <andresvi@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>