[fidl][lint] Support enabling  experimental checks
Add a new --experimental-checks flag to the linter to allow enabling
checks without switching the linter entirely into exclude-by-default
mode. Also add support for setting this new flag to the FIDL build rule.

Also change the existing "no-trailing-comment" check to use the
experimental-checks mechanism. It was previously explicitly added to the
excluded checks list to disable it, but with the experimental checks
mechanism, the same thing can be supported in a more explicit way. This
is backwards compatible since the experimental check can still be
enabled by the --include-checks flag, which is the only way someone
could previously have been using it.

Bug: 88366
Change-Id: Ie62f0505d39ddba832cbe412741dbaf5266a2391
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762868
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Zachary Stewart <zstewart@google.com>